# Optimize buffer size for static file reads