## Description
<!--- Describe your changes but trying to be concise as possible -->

This PR optimizes the buffer size (adapting the code again) when reading files using the OS `blocksize` when available.
As a result, it reduces the RAM utilization by around 15% and increases slightly the req/sec average.

All Linux/Unix targets should be affected positively by this improvement.

Note: this was inspired by [warp `fs`](https://github.com/seanmonstar/warp/). 

<!--- Provide a general summary of your changes in the Title above -->

## Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
<!--- If it fixes an open issue, please link to the issue here. -->

Our effort to improve SWS performance.

## How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->

## Screenshots (if appropriate):

**Linux x86_64 (dynamic binary)**

- Intel Core i7 8GB
- Linux 6.5.5-arch1-1
- Rust 1.73.0

--

Before optimization (v2.22.1)

```sh
# With static buffer (default 4KB) - ~3.2MB RAM (max)
$ wrk -c100 -t4 -d1m --latency http://localhost:8787
# Running 1m test @ http://localhost:8787
#   4 threads and 100 connections
#   Thread Stats   Avg      Stdev     Max   +/- Stdev
#     Latency     2.29ms    0.92ms  23.99ms   72.48%
#     Req/Sec    11.00k   553.68    13.89k    83.38%
#   Latency Distribution
#      50%    2.25ms
#      75%    2.81ms
#      90%    3.39ms
#      99%    4.70ms
#   2625903 requests in 1.00m, 1.88GB read
# Requests/sec:  43741.67
# Transfer/sec:     32.00MB
```

After optimization

```sh
# With optimal buffer (default 4KB) - ~2.7MB RAM (max)
wrk -c100 -t4 -d1m --latency http://localhost:8787
# Running 1m test @ http://localhost:8787
#   4 threads and 100 connections
#   Thread Stats   Avg      Stdev     Max   +/- Stdev
#     Latency     2.27ms    0.91ms  20.33ms   72.38%
#     Req/Sec    11.09k   573.89    15.13k    83.54%
#   Latency Distribution
#      50%    2.23ms
#      75%    2.79ms
#      90%    3.37ms
#      99%    4.69ms
#   2648912 requests in 1.00m, 1.89GB read
# Requests/sec:  44110.94
# Transfer/sec:     32.27MB
```

**~15.6%** less RAM utilization.

More details at https://gist.github.com/joseluisq/0b18c10d89fdf01beeb9e9b9b03cb761